admin/backend/requirements.txt
# FastAPI Admin Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Cache
redis[hiredis]==5.0.1

# Security & Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client (for external API calls)
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1